requirements.txt
pytest
pandas
numpy
scikit-learn
lightgbm
optuna
duckdb
pyyaml
pydantic
black
isort
pylint